+++ vs-admin/src/gui.rs
@@ -6,8 +6,8 @@ use ratatui::{
     style::{Color, Style, Stylize},
     text::{Line, Span, Text},
     widgets::{
-        Block, BorderType, Borders, Cell, Paragraph, Row, Scrollbar, ScrollbarOrientation,
-        ScrollbarState, Table, TableState,
+        Block, Cell, Paragraph, Row, Scrollbar, ScrollbarOrientation, ScrollbarState, Table,
+        TableState,
     },
 };
 use reqwest::tls::Certificate;
@@ -16,45 +16,24 @@ use chrono::{DateTime, SecondsFormat, Utc};
 use std::collections::VecDeque;
 use std::time::{Duration, Instant, SystemTime};
 
-use admin_api_types::{
-    ActorDescriptor, ApiKeySet, NodeRecordBrief, ServiceDescriptor, VisaDescriptor,
-};
-
-use thousands::Separable;
-use unicode_width::{UnicodeWidthChar, UnicodeWidthStr};
+use admin_api_types::{ActorDescriptor, ServiceDescriptor, VisaDescriptor};
 
 use crate::vsclient::{RoleFilter, VsClient};
 
+mod details;
+mod format;
+mod text;
+mod widgets;
+
+use details::{actor_detail_text, service_detail_text, visa_detail_text};
+use format::fmt_uptime;
+use widgets::{
+    SPARK_MAX, SPARK_MIN, clamp_scroll, clamp_state, detail_line, move_selection, pane_block,
+    sparkline,
+};
+
 /// Do not hit the VS ADMIN api more than this often.
 const REFRESH_RATE: Duration = Duration::from_millis(2000);
-
-/// Minimum sparkline width (bars) when shown; below this the sparkline is dropped.
-const SPARK_MIN: usize = 8;
-/// Maximum sparkline width (bars); history ring buffers hold this many samples.
-const SPARK_MAX: usize = 16;
-
-/// Render `vals` as unicode bars scaled to the slice max; all-min when idle.
-fn sparkline(vals: &[u64]) -> String {
-    const BARS: [char; 8] = ['▁', '▂', '▃', '▄', '▅', '▆', '▇', '█'];
-    let max = vals.iter().copied().max().unwrap_or(0);
-    if max == 0 {
-        return "▁".repeat(vals.len());
-    }
-    vals.iter()
-        .map(|&v| BARS[((v * 7) / max) as usize])
-        .collect()
-}
-
-/// Format seconds as "01d 04h 09m 22s", hours rolling into days.
-fn fmt_uptime(secs: u64) -> String {
-    format!(
-        "{:02}d {:02}h {:02}m {:02}s",
-        secs / 86400,
-        secs % 86400 / 3600,
-        secs % 3600 / 60,
-        secs % 60,
-    )
-}
 
 /// Which content pane is currently selected for keyboard interaction.
 #[derive(Debug, Clone, Copy, PartialEq)]
@@ -63,501 +42,6 @@ enum Pane {
     Services,
     Visas,
     Details,
-}
-
-/// Clamp a scroll offset so the last content line can reach the pane bottom
-/// but you can't scroll into empty space; 0 when content fits the viewport.
-fn clamp_scroll(offset: u16, content_h: u16, viewport_h: u16) -> u16 {
-    offset.min(content_h.saturating_sub(viewport_h))
-}
-
-/// Next selection index given current index, list length, and direction.
-/// Clamps at 0 and len-1; None only for an empty list.
-fn move_selection(current: Option<usize>, len: usize, down: bool) -> Option<usize> {
-    if len == 0 {
-        return None;
-    }
-    let cur = current.unwrap_or(0);
-    Some(if down {
-        (cur + 1).min(len - 1)
-    } else {
-        cur.saturating_sub(1)
-    })
-}
-
-/// One-line description of the entity feeding the details pane, e.g.
-/// "visa 1001" or "actor (none selected)".
-fn detail_line(source: Pane, id: Option<String>) -> String {
-    let label = match source {
-        Pane::Actors => "actor",
-        Pane::Services => "service",
-        Pane::Visas => "visa",
-        Pane::Details => return String::new(),
-    };
-    match id {
-        Some(id) => format!("{label} {id}"),
-        None => format!("{label} (none selected)"),
-    }
-}
-
-/// Width of the label column in the visa Details view. 13 = longest label
-/// ("last contact"/"connect reqs", 12 chars) plus a one-space gutter.
-const LABEL_W: u16 = 13;
-
-/// Color for the labels/headings in the visa Details view (vs. plain values).
-const HEADING_STYLE: Style = Style::new().fg(Color::Cyan);
-
-/// Attribute key / value colors (distinct from each other and the heading).
-const ATTR_KEY_STYLE: Style = Style::new().fg(Color::Yellow);
-const ATTR_VAL_STYLE: Style = Style::new().fg(Color::White);
-
-/// "Xh Ym" for a minute count, hours comma-separated (durations can be huge).
-fn hm(total_mins: u64) -> String {
-    format!(
-        "{}h {:02}m",
-        (total_mins / 60).separate_with_commas(),
-        total_mins % 60
-    )
-}
-
-/// Human-readable time remaining until `expires`, relative to `now`.
-/// "EXPIRED" once past; otherwise "expires in Xh Ym" (or "expires in Xm"
-/// under an hour). Rounds down.
-fn remaining_str(expires: SystemTime, now: SystemTime) -> String {
-    match expires.duration_since(now) {
-        Err(_) => "EXPIRED".to_string(),
-        Ok(d) => {
-            let mins = d.as_secs() / 60;
-            if mins >= 60 {
-                format!("expires in {}", hm(mins))
-            } else {
-                format!("expires in {mins}m")
-            }
-        }
-    }
-}
-
-/// The flow line: `[src]:sport  --PROTO-->  [dst]:dport`. Missing addr → `-`,
-/// missing port → `0` (matches the visas list row convention).
-fn flow_str(v: &VisaDescriptor) -> String {
-    let src = v.source_addr.as_deref().unwrap_or("-");
-    let dst = v.dest_addr.as_deref().unwrap_or("-");
-    let sport = v.source_port.unwrap_or(0);
-    let dport = v.dest_port.unwrap_or(0);
-    let proto = v.proto.to_uppercase();
-    format!("[{src}]:{sport}  --{proto}-->  [{dst}]:{dport}")
-}
-
-/// One-line session-key summary — never prints key material, only the key
-/// format and the encrypted byte lengths.
-fn session_key_summary(k: &ApiKeySet) -> String {
-    format!(
-        "fmt={:?} ingress={}B egress={}B",
-        k.format,
-        k.ingress_key.len(),
-        k.egress_key.len()
-    )
-}
-
-/// A labeled row: `label` (heading color) in a fixed `LABEL_W` column, `value`
-/// wrapped to the remaining width with continuation lines indented under the
-/// value column.
-fn labeled(label: &str, value: &str, width: u16) -> Vec<Line<'static>> {
-    let lw = LABEL_W as usize;
-    // Column available for the value; keep at least 1 so we always progress.
-    let vw = (width.saturating_sub(LABEL_W)).max(1) as usize;
-    let wrapped = wrap_words(value, vw);
-    let mut lines = Vec::with_capacity(wrapped.len().max(1));
-    for (i, seg) in wrapped.iter().enumerate() {
-        let label_span = if i == 0 {
-            Span::from(format!("{label:<lw$}")).style(HEADING_STYLE)
-        } else {
-            Span::from(" ".repeat(lw))
-        };
-        lines.push(Line::from(vec![label_span, Span::from(seg.clone())]));
-    }
-    lines
-}
-
-/// Word-wrap `s` to `width` cols. Splits on spaces; a single word longer than
-/// `width` is hard-split. Always returns at least one segment.
-fn wrap_words(s: &str, width: usize) -> Vec<String> {
-    let width = width.max(1);
-    let mut out = Vec::new();
-    let mut cur = String::new();
-    for word in s.split_whitespace() {
-        let word_w = UnicodeWidthStr::width(word);
-        if word_w > width {
-            if !cur.is_empty() {
-                out.push(std::mem::take(&mut cur));
-            }
-            let chunks = split_word_by_display_width(word, width);
-            let last = chunks.len().saturating_sub(1);
-            for (i, chunk) in chunks.into_iter().enumerate() {
-                if i == last {
-                    cur = chunk;
-                } else {
-                    out.push(chunk);
-                }
-            }
-            continue;
-        }
-        if cur.is_empty() {
-            cur = word.to_string();
-        } else if UnicodeWidthStr::width(cur.as_str()) + 1 + word_w <= width {
-            cur.push(' ');
-            cur.push_str(word);
-        } else {
-            out.push(std::mem::take(&mut cur));
-            cur = word.to_string();
-        }
-    }
-    out.push(cur);
-    out
-}
-
-/// Split one whitespace-free string into UTF-8-safe display-width chunks.
-fn split_word_by_display_width(s: &str, width: usize) -> Vec<String> {
-    let width = width.max(1);
-    let mut chunks = Vec::new();
-    let mut cur = String::new();
-    let mut cur_w = 0;
-
-    for ch in s.chars() {
-        let ch_w = UnicodeWidthChar::width(ch).unwrap_or(0).max(1);
-        if cur_w > 0 && cur_w + ch_w > width {
-            chunks.push(std::mem::take(&mut cur));
-            cur_w = 0;
-        }
-        cur.push(ch);
-        cur_w += ch_w;
-    }
-
-    if !cur.is_empty() {
-        chunks.push(cur);
-    }
-    chunks
-}
-
-/// Build the full multi-line visa detail view for the Details pane.
-/// Pure: no `self`, no frame — `width` is the inner (border-excluded) width.
-fn visa_detail_text(
-    v: &VisaDescriptor,
-    cn: Option<&str>,
-    now: SystemTime,
-    width: u16,
-) -> Text<'static> {
-    let w = width as usize;
-    let mut lines: Vec<Line> = Vec::new();
-
-    // Header: "visa <id>" left, remaining right-aligned, padded to width.
-    let left = format!("visa {}", v.id);
-    let right = remaining_str(v.expires, now);
-    let pad = w.saturating_sub(left.len() + right.len());
-    lines.push(Line::from(vec![
-        Span::from("visa ").style(HEADING_STYLE),
-        Span::from(v.id.to_string()),
-        Span::from(" ".repeat(pad)),
-        Span::from(right).style(HEADING_STYLE),
-    ]));
-    lines.push(Line::from(""));
-
-    lines.extend(labeled("policy", &v.policy_id, width));
-    lines.extend(labeled(
-        "zpl",
-        &format!("[{}] {}", v.direction, v.zpl),
-        width,
-    ));
-    let requesting = match cn {
-        Some(cn) => format!("{} ({cn})", v.requesting_node),
-        None => v.requesting_node.clone(),
-    };
-    lines.extend(labeled("requesting", &requesting, width));
-    lines.extend(labeled("flow", &flow_str(v), width));
-
-    let created: DateTime<Utc> = v.created.into();
-    let expires: DateTime<Utc> = v.expires.into();
-    lines.extend(labeled(
-        "created",
-        &created.to_rfc3339_opts(SecondsFormat::Secs, true),
-        width,
-    ));
-    let dur = v
-        .expires
-        .duration_since(v.created)
-        .map(|d| hm(d.as_secs() / 60))
-        .unwrap_or_else(|_| "0h 00m".to_string());
-    lines.extend(labeled(
-        "expires",
-        &format!(
-            "{}  (duration {dur})",
-            expires.to_rfc3339_opts(SecondsFormat::Secs, true)
-        ),
-        width,
-    ));
-
-    // Signals: first on the label line, the rest indented under it.
-    if v.signals.is_empty() {
-        lines.extend(labeled("signals", "(none)", width));
-    } else {
-        for (i, sig) in v.signals.iter().enumerate() {
-            let label = if i == 0 { "signals" } else { "" };
-            lines.extend(labeled(label, sig, width));
-        }
-    }
-
-    lines.extend(labeled(
-        "session key",
-        &session_key_summary(&v.session_key),
-        width,
-    ));
-
-    Text::from(lines)
-}
-
-/// A `SystemTime` as RFC3339 (secs, UTC), or `fallback` when absent.
-fn ts_or(t: Option<SystemTime>, fallback: &str) -> String {
-    match t {
-        Some(t) => {
-            let dt: DateTime<Utc> = t.into();
-            dt.to_rfc3339_opts(SecondsFormat::Secs, true)
-        }
-        None => fallback.to_string(),
-    }
-}
-
-/// Header's right-hand auth text: "no auth" (None), "auth EXPIRED" (past),
-/// or "auth in Xh Ym".
-fn auth_hdr(auth_exp: Option<SystemTime>, now: SystemTime) -> String {
-    match auth_exp {
-        None => "no auth".to_string(),
-        Some(e) => match e.duration_since(now) {
-            Err(_) => "auth EXPIRED".to_string(),
-            Ok(d) => format!("auth in {}", hm(d.as_secs() / 60)),
-        },
-    }
-}
-
-/// Relative " (Xm ago)" / " (Xh Ym ago)" suffix for a past time; empty if the
-/// time is in the future or absent.
-fn ago_suffix(t: SystemTime, now: SystemTime) -> String {
-    match now.duration_since(t) {
-        Ok(d) => {
-            let mins = d.as_secs() / 60;
-            if mins >= 60 {
-                format!(" ({} ago)", hm(mins))
-            } else {
-                format!(" ({mins}m ago)")
-            }
-        }
-        Err(_) => String::new(),
-    }
-}
-
-/// The `-- node --` block for a node actor's `NodeRecordBrief`.
-/// `width` is the inner (border-excluded) pane width.
-fn node_detail_lines(nd: &NodeRecordBrief, now: SystemTime, width: u16) -> Vec<Line<'static>> {
-    let mut lines: Vec<Line> = Vec::new();
-
-    // Full-width "-- node ----" divider.
-    let mut divider = String::from("-- node ");
-    let pad = (width as usize).saturating_sub(divider.len());
-    divider.push_str(&"-".repeat(pad));
-    lines.push(Line::from(Span::from(divider).style(HEADING_STYLE)));
-
-    lines.extend(labeled(
-        "sync",
-        if nd.in_sync { "YES" } else { "NO" },
-        width,
-    ));
-    let last_contact = format!(
-        "{}{}",
-        ts_or(nd.last_contact, "never"),
-        nd.last_contact
-            .map(|t| ago_suffix(t, now))
-            .unwrap_or_default(),
-    );
-    lines.extend(labeled("last contact", &last_contact, width));
-    lines.extend(labeled(
-        "vss port",
-        &nd.vss_port.map(|p| p.to_string()).unwrap_or("-".into()),
-        width,
-    ));
-    lines.extend(labeled(
-        "visa reqs",
-        &format!(
-            "{}  (approved {}, denied {})",
-            nd.visa_requests.separate_with_commas(),
-            nd.approved_vreqs.separate_with_commas(),
-            nd.denied_vreqs.separate_with_commas(),
-        ),
-        width,
-    ));
-    let last_vreq = format!(
-        "{}{}",
-        ts_or(nd.last_vreq, "never"),
-        nd.last_vreq.map(|t| ago_suffix(t, now)).unwrap_or_default(),
-    );
-    lines.extend(labeled("last vreq", &last_vreq, width));
-    lines.extend(labeled(
-        "connect reqs",
-        &nd.connect_requests.separate_with_commas(),
-        width,
-    ));
-    // installed = live visas, pending = awaiting install, revoking = pending
-    // revocation. `visas_enqueued` (queued IDs) is not surfaced here.
-    lines.extend(labeled(
-        "installs",
-        &format!(
-            "{} installed, {} pending, {} revoking",
-            nd.visas.len().separate_with_commas(),
-            nd.pending_install.separate_with_commas(),
-            nd.pending_revocation.separate_with_commas(),
-        ),
-        width,
-    ));
-    let adapters = if nd.adapters.is_empty() {
-        "(none)".to_string()
-    } else {
-        nd.adapters.join(", ")
-    };
-    lines.extend(labeled("adapters", &adapters, width));
-    let links = if nd.links.is_empty() {
-        "(none)".to_string()
-    } else {
-        nd.links.join(", ")
-    };
-    lines.extend(labeled("links", &links, width));
-
-    lines
-}
-
-/// Build the full multi-line actor detail view for the Details pane.
-/// `services` is the CNs' bound service names (already filtered by caller).
-/// Pure: no `self`, no frame — `width` is the inner (border-excluded) width.
-fn actor_detail_text(
-    a: &ActorDescriptor,
-    services: &[&str],
-    now: SystemTime,
-    width: u16,
-) -> Text<'static> {
-    let w = width as usize;
-    let mut lines: Vec<Line> = Vec::new();
-
-    // Header: "actor <cn>" left; "[node]/[adapter]   <auth>" right, padded.
-    let badge = if a.node { "[node]" } else { "[adapter]" };
-    let right = format!("{badge}   {}", auth_hdr(a.auth_exp, now));
-    let left = format!("actor {}", a.cn);
-    let pad = w.saturating_sub(left.len() + right.len());
-    lines.push(Line::from(vec![
-        Span::from("actor ").style(HEADING_STYLE),
-        Span::from(a.cn.clone()),
-        Span::from(" ".repeat(pad)),
-        Span::from(right).style(HEADING_STYLE),
-    ]));
-    lines.push(Line::from(""));
-
-    lines.extend(labeled("identity", &a.ident, width));
-    lines.extend(labeled("zpr addr", &a.zpr_addr, width));
-    let created: DateTime<Utc> = a.ctime.into();
-    lines.extend(labeled(
-        "created",
-        &created.to_rfc3339_opts(SecondsFormat::Secs, true),
-        width,
-    ));
-    // auth exp: timestamp plus a relative "(in Xh Ym)"/"(EXPIRED)" suffix.
-    let auth_suffix = match a.auth_exp {
-        Some(e) => match e.duration_since(now) {
-            Err(_) => "  (EXPIRED)".to_string(),
-            Ok(d) => format!("  (in {})", hm(d.as_secs() / 60)),
-        },
-        None => String::new(),
-    };
-    lines.extend(labeled(
-        "auth exp",
-        &format!("{}{auth_suffix}", ts_or(a.auth_exp, "no auth")),
-        width,
-    ));
-
-    // attributes: first "key=values" on the label line, rest indented.
-    // Sort by key so the order is stable across refreshes (attrs arrive unordered).
-    if a.attrs.is_empty() {
-        lines.extend(labeled("attributes", "(none)", width));
-    } else {
-        let mut attrs: Vec<&_> = a.attrs.iter().collect();
-        attrs.sort_by(|x, y| x.key.cmp(&y.key));
-        let lw = LABEL_W as usize;
-        for (i, attr) in attrs.iter().enumerate() {
-            let label = if i == 0 { "attributes" } else { "" };
-            // Colored key = value; one line per attribute.
-            // ponytail: no wrap — attr values are short; long ones clip at the
-            // right edge (vertical scroll won't help). Widen if that bites.
-            lines.push(Line::from(vec![
-                Span::from(format!("{label:<lw$}")).style(HEADING_STYLE),
-                Span::from(attr.key.clone()).style(ATTR_KEY_STYLE),
-                Span::from(" = "),
-                Span::from(attr.value.join(", ")).style(ATTR_VAL_STYLE),
-            ]));
-        }
-    }
-
-    // services: one bound service name per line, or (none).
-    if services.is_empty() {
-        lines.extend(labeled("services", "(none)", width));
-    } else {
-        for (i, s) in services.iter().enumerate() {
-            let label = if i == 0 { "services" } else { "" };
-            lines.extend(labeled(label, s, width));
-        }
-    }
-
-    // node block only for node actors.
-    if let Some(nd) = &a.node_details {
-        lines.extend(node_detail_lines(nd, now, width));
-    }
-
-    Text::from(lines)
-}
-
-/// Full multi-line detail view for a selected service. Header is the service
-/// name; body is one labeled row per field. All fields are plain strings.
-fn service_detail_text(s: &ServiceDescriptor, width: u16) -> Text<'static> {
-    let mut lines: Vec<Line> = Vec::new();
-    lines.push(Line::from(vec![
-        Span::from("service ").style(HEADING_STYLE),
-        Span::from(s.service_name.clone()),
-    ]));
-    lines.push(Line::from(""));
-    lines.extend(labeled("cn", &s.actor_cn, width));
-    lines.extend(labeled("zpr addr", &s.zpr_addr, width));
-    lines.extend(labeled("dock addr", &s.dock_zpr_addr, width));
-    lines.extend(labeled("kind", &s.service_kind, width));
-    lines.extend(labeled("endpoints", &s.service_endpoints, width));
-    Text::from(lines)
-}
-
-/// Keep a stored selection in range after the row vec is rebuilt.
-fn clamp_state(state: &mut TableState, len: usize) {
-    state.select(match (state.selected(), len) {
-        (_, 0) => None,
-        (Some(i), _) if i >= len => Some(len - 1),
-        (None, _) => Some(0),
-        (some, _) => some,
-    });
-}
-
-/// Bordered block for a pane; thick bold border when selected, and the
-/// hot-key letter of the title underlined+bold.
-fn pane_block<'a>(letter: &'a str, rest: &'a str, selected: bool) -> Block<'a> {
-    let title = Line::from(vec![letter.underlined().bold(), rest.bold()]);
-    let block = Block::default().borders(Borders::ALL).title(title);
-    if selected {
-        block
-            .border_type(BorderType::Thick)
-            .border_style(Style::default().bold())
-    } else {
-        block
-    }
 }
 
 #[derive(Debug)]
@@ -1257,265 +741,5 @@ impl Gui {
         let text = detail_line(self.detail_source, id);
         // Placeholder is short, so max == 0 → no scrollbar; same path is harmless.
         self.render_detail_text(frame, area, Text::from(text));
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        LABEL_W, Pane, actor_detail_text, auth_hdr, clamp_scroll, detail_line, flow_str,
-        fmt_uptime, labeled, move_selection, remaining_str, service_detail_text,
-        session_key_summary, sparkline, ts_or, wrap_words,
-    };
-    use admin_api_types::{
-        ActorDescriptor, ApiAttribute, ApiKeySet, NodeRecordBrief, ServiceDescriptor,
-        VisaDescriptor, VisaMatchDirection,
-    };
-    use std::time::{Duration, UNIX_EPOCH};
-
-    /// A minimal ActorDescriptor; `node_details`/`attrs` set per-test.
-    fn sample_actor() -> ActorDescriptor {
-        ActorDescriptor {
-            cn: "web-server-01".into(),
-            ctime: UNIX_EPOCH,
-            ident: "CN=web-server-01,O=ZPR".into(),
-            node: false,
-            zpr_addr: "fd5a:5052::b2".into(),
-            attrs: vec![],
-            auth_exp: None,
-            node_details: None,
-        }
-    }
-
-    /// A NodeRecordBrief fixture for the "-- node --" block.
-    fn sample_node() -> NodeRecordBrief {
-        NodeRecordBrief {
-            pending_install: 3,
-            last_contact: Some(UNIX_EPOCH + Duration::from_secs(600)),
-            visa_requests: 1204,
-            connect_requests: 5309,
-            in_sync: true,
-            approved_vreqs: 1180,
-            denied_vreqs: 24,
-            last_vreq: Some(UNIX_EPOCH + Duration::from_secs(500)),
-            adapters: vec!["adapter-a".into(), "adapter-b".into()],
-            links: vec!["node-2".into()],
-            visas: vec![1, 2, 3],
-            visas_enqueued: vec![4],
-            pending_revocation: 0,
-            vss_port: Some(8443),
-        }
-    }
-
-    /// Join a Text's lines into one string for substring assertions.
-    fn text_str(t: &super::Text<'static>) -> String {
-        t.lines
-            .iter()
-            .map(|l| l.to_string())
-            .collect::<Vec<_>>()
-            .join("\n")
-    }
-
-    /// A minimal VisaDescriptor for helper tests.
-    fn sample_visa() -> VisaDescriptor {
-        VisaDescriptor {
-            id: 1001,
-            expires: UNIX_EPOCH + Duration::from_secs(3600),
-            created: UNIX_EPOCH,
-            policy_id: "v7".into(),
-            zpl: "allow tcp alice -> web:443".into(),
-            direction: VisaMatchDirection::Forward,
-            requesting_node: "fd5a:5052::a1".into(),
-            source_addr: Some("fd5a:5052::a1".into()),
-            dest_addr: Some("fd5a:5052::b2".into()),
-            source_port: Some(52344),
-            dest_port: Some(443),
-            proto: "tcp".into(),
-            signals: vec![],
-            session_key: ApiKeySet {
-                format: Default::default(),
-                ingress_key: vec![0u8; 48],
-                egress_key: vec![0u8; 48],
-            },
-        }
-    }
-
-    /// move_selection clamps at both ends and handles the empty list.
-    #[test]
-    fn move_selection_clamps_and_steps() {
-        assert_eq!(move_selection(None, 0, true), None);
-        assert_eq!(move_selection(Some(0), 3, false), Some(0)); // clamp top
-        assert_eq!(move_selection(Some(2), 3, true), Some(2)); // clamp bottom
-        assert_eq!(move_selection(Some(0), 3, true), Some(1)); // step down
-        assert_eq!(move_selection(Some(2), 3, false), Some(1)); // step up
-    }
-
-    /// detail_line formats "<entity> <id>" and falls back when nothing selected.
-    #[test]
-    fn detail_line_formats_and_falls_back() {
-        assert_eq!(detail_line(Pane::Visas, Some("1001".into())), "visa 1001");
-        assert_eq!(detail_line(Pane::Actors, None), "actor (none selected)");
-    }
-
-    /// remaining_str shows EXPIRED past, and hh/mm vs minutes-only.
-    #[test]
-    fn remaining_str_formats_and_expires() {
-        let base = UNIX_EPOCH;
-        // 1h02m ahead.
-        let exp = base + Duration::from_secs(3720);
-        assert_eq!(remaining_str(exp, base), "expires in 1h 02m");
-        // 45m ahead → minutes only.
-        let exp = base + Duration::from_secs(45 * 60);
-        assert_eq!(remaining_str(exp, base), "expires in 45m");
-        // already past.
-        assert_eq!(
-            remaining_str(base, base + Duration::from_secs(1)),
-            "EXPIRED"
-        );
-    }
-
-    /// flow_str renders full flows and uses -/0 for missing addr/port.
-    #[test]
-    fn flow_str_full_and_missing() {
-        let v = sample_visa();
-        assert_eq!(
-            flow_str(&v),
-            "[fd5a:5052::a1]:52344  --TCP-->  [fd5a:5052::b2]:443"
-        );
-        let mut v = sample_visa();
-        v.source_addr = None;
-        v.dest_port = None;
-        assert_eq!(flow_str(&v), "[-]:52344  --TCP-->  [fd5a:5052::b2]:0");
-    }
-
-    /// session_key_summary reports format and byte lengths, never key bytes.
-    #[test]
-    fn session_key_summary_no_material() {
-        let s = session_key_summary(&sample_visa().session_key);
-        assert!(s.contains("fmt="));
-        assert!(s.contains("ingress=48B"));
-        assert!(s.contains("egress=48B"));
-    }
-
-    /// labeled wraps long values and indents continuation lines by LABEL_W.
-    #[test]
-    fn labeled_wraps_and_indents() {
-        let lines = labeled("zpl", "alpha bravo charlie delta echo foxtrot", 24);
-        assert!(lines.len() > 1, "expected wrapping");
-        // Continuation line begins with LABEL_W spaces.
-        let second = lines[1].to_string();
-        assert!(second.starts_with(&" ".repeat(LABEL_W as usize)));
-    }
-
-    /// wrap_words hard-splits non-ASCII text without cutting UTF-8 bytes.
-    #[test]
-    fn wrap_words_handles_unicode_hard_split() {
-        assert_eq!(wrap_words("éé", 1), vec!["é".to_string(), "é".to_string()]);
-        assert_eq!(
-            wrap_words("日本語", 2),
-            vec!["日".to_string(), "本".to_string(), "語".to_string()]
-        );
-    }
-
-    /// auth_hdr covers None, expired, and future cases.
-    #[test]
-    fn auth_hdr_none_expired_future() {
-        let base = UNIX_EPOCH;
-        assert_eq!(auth_hdr(None, base), "no auth");
-        // 1s in the past → EXPIRED.
-        assert_eq!(
-            auth_hdr(Some(base), base + Duration::from_secs(1)),
-            "auth EXPIRED"
-        );
-        // 1h02m ahead.
-        let exp = base + Duration::from_secs(3720);
-        assert_eq!(auth_hdr(Some(exp), base), "auth in 1h 02m");
-    }
-
-    /// ts_or renders RFC3339 when set, else the fallback word.
-    #[test]
-    fn ts_or_timestamp_or_fallback() {
-        assert!(ts_or(Some(UNIX_EPOCH), "never").contains("1970-01-01"));
-        assert_eq!(ts_or(None, "never"), "never");
-    }
-
-    /// A node actor shows the badge, the node block, and its bound services.
-    #[test]
-    fn actor_detail_text_node() {
-        let mut a = sample_actor();
-        a.node = true;
-        a.attrs = vec![ApiAttribute {
-            key: "role".into(),
-            value: vec!["node".into()],
-            expires_at: UNIX_EPOCH,
-        }];
-        a.node_details = Some(sample_node());
-        let s = text_str(&actor_detail_text(&a, &["https-web"], UNIX_EPOCH, 78));
-        assert!(s.contains("[node]"));
-        assert!(s.contains("-- node"));
-        assert!(s.contains("sync"));
-        assert!(s.contains("role = node"));
-        assert!(s.contains("https-web"));
-    }
-
-    /// An adapter (no node_details) omits the node block; empty attrs/services
-    /// render "(none)".
-    #[test]
-    fn actor_detail_text_adapter_empty() {
-        let a = sample_actor(); // node=false, no details/attrs
-        let s = text_str(&actor_detail_text(&a, &[], UNIX_EPOCH, 78));
-        assert!(s.contains("[adapter]"));
-        assert!(!s.contains("-- node"));
-        assert!(s.contains("attributes   (none)"));
-        assert!(s.contains("services     (none)"));
-    }
-
-    /// service_detail_text: header carries the service name and every field
-    /// value shows up in the rendered lines.
-    #[test]
-    fn service_detail_text_fields() {
-        let sd = ServiceDescriptor {
-            service_name: "webproxy".into(),
-            actor_cn: "actor-7".into(),
-            zpr_addr: "fd5a:5052::1".into(),
-            dock_zpr_addr: "fd5a:5052::2".into(),
-            service_kind: "https".into(),
-            service_endpoints: "https://host:443".into(),
-        };
-        let s = text_str(&service_detail_text(&sd, 78));
-        assert!(s.contains("service webproxy"));
-        assert!(s.contains("actor-7"));
-        assert!(s.contains("fd5a:5052::1"));
-        assert!(s.contains("fd5a:5052::2"));
-        assert!(s.contains("https"));
-        assert!(s.contains("https://host:443"));
-    }
-
-    /// clamp_scroll: fits-in-view clamps to 0; past-end clamps to the max;
-    /// in-range is untouched.
-    #[test]
-    fn clamp_scroll_bounds() {
-        // Content shorter than viewport → no scroll.
-        assert_eq!(clamp_scroll(5, 10, 20), 0);
-        // Offset past the end → content_h - viewport_h.
-        assert_eq!(clamp_scroll(100, 30, 10), 20);
-        // Offset within range → unchanged.
-        assert_eq!(clamp_scroll(3, 30, 10), 3);
-        // Exactly fits → 0.
-        assert_eq!(clamp_scroll(4, 10, 10), 0);
-    }
-
-    /// fmt_uptime rolls seconds into d/h/m/s with zero-padding.
-    #[test]
-    fn fmt_uptime_rolls_units() {
-        assert_eq!(fmt_uptime(100162), "01d 03h 49m 22s");
-        assert_eq!(fmt_uptime(0), "00d 00h 00m 00s");
-    }
-
-    /// sparkline: no activity is all-min; a monotonic rise maxes at the top bar.
-    #[test]
-    fn sparkline_idle_and_peak() {
-        assert_eq!(sparkline(&[0, 0]), "▁▁");
-        assert!(sparkline(&[1, 2, 3]).ends_with('█'));
     }
 }
